+++ Code/glimac/src/project/StaticObject.cpp
@@ -1,7 +1,7 @@
 #include "../include/project/StaticObject.hpp"
 
 
-StaticObject::StaticObject(char type, int posX, int posY, float width, float height, Orientation orientation) : Object(type, posX, posY, width, height, orientation) { }
+StaticObject::StaticObject(char type, float posX, float posY, float width, float height, Orientation orientation) : Object(type, posX, posY, width, height, orientation) { }
 StaticObject::StaticObject() { }
