+++ Code/glimac/src/project/Wall.cpp
@@ -1,7 +1,7 @@
 #include "project/Wall.hpp"
 
 // Constructors
-Wall::Wall(int posX, int posY, float width, float height, Orientation orientation) :
+Wall::Wall(float posX, float posY, float width, float height, Orientation orientation) :
     StaticObject('W', posX, posY, width, height, orientation) { }
 Wall::Wall() { }
